app/src/backend/requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
flask-cors
WTForms
psycopg2
pycryptodome